crypto: openssl: Remove email address in DES benchmark
Note from GitHub support:

I'm reaching out because we've been contacted by a person who has requested removal of potentially personal information in one of the repositories you own. Here is the specific content they identified via a data subject request:

https://github.com/kdrag0n/netbsd_easel/blob/trunk/crypto/external/bsd/openssl.old/dist/crypto/des/times/aix.cc

The line that should be removed in all of them is number 1, which contains an email address.

They are unable to remove the content themselves since it is in a repository that you own. As a processor, GitHub is notifying you and passing the removal request along to you. As the controller of that data, it is your decision how to handle this request.

Please let me know if you have any further questions, and thanks for your attention.